refactor(parse): move helper fuctions to its own file
