+++ .changeset/fuzzy-ghosts-battle.md
@@ -1,0 +1,5 @@
+---
+'@lottiefiles/relottie-parse': patch
+---
+
+refactor(parse): move traverse assistent logic to helpers & Stack to its own file
